Continuing the shader loading logic (now makefile)
We now have a osl/ directory that contains Houdini shaders we want to
implement. The target is to have a good amount of VEX shaders ported in
there. There is an empty hlight.osl shader in there for cmake testing
purposes. The compiled shaders will be put in:

	path_to_rop/osl/

Note that from now on one has to run the 'install' (e.g. 'ninja
install') as OSL shaders are birst compiled to build/osl/ and then
copied to $HIH/osl.