Work in progress for NSI light shader.
We consider the OBJ_Node of the light as being a VOP shader for the
purpose of the render. It derives from an OP_Parameters node so we can
use the same functionality. We have a 'hlight.osl' shader which dictates
what parameters in the OP_Parameters we actually need to export (there
are many parameters in the Houdini light that are not needed for the
shader).

To enumerate the parameters of hilight.osl we use the newly introducted
DlShaderInfo functionality so a recent version of 3Delight is necessary
after this commit.