Make it light and make it shade mofos!
On the lighting side, our support of Houdini's light is getting good.
Shading logic is in "hlight.osl" and geometry export logic is in
light.cpp. Not all the light types are tested but one can play with
area lights and even geo lights. We will have more work to do
but for now this will allow us to get some good progress on other
things.

On the shading side things are more interesting: we can now parse
networks in /mat and have them assigned to objects. Our way of doing
things is simple: we try to find an OSL shader corresponding to the
every shader node we encounter. For example, if we encounter
principledshader::2.0 et will look for:

	installationpath/osl/principledshader__2_0.oso

We will then parse all the parameters of that OSL shader and
follow any potential links for these supported parameters. This
will allow us to add support for VEX shaders progressively. One
tricky part is Houdini shaders that have subnetworks. These can
be supported in two ways:

1 - Implement every shader in the subnetwork.
2 - Implement a monolithic shader node the does what the subnetwork
    does. This will be the only way for shaders which contain
    inline code.

An example for #2 is texture::2.0. It contains code, namely for
handling UDIMs, that is inlined and not convertible. We have a
simple mechanism for that: use the "ignoresubnetworks" tag in
the OSL shader so that 3Delight for Houdini doesn't try to dig
into the subnetworks. This is still WIP but gives you an idea
of the logic.

I have added simple placeholders in /osl for both the principled and the
texture shader nodes.