Implement spot light functionality.
We don't do this the same way as Houdini. Houdini will allow you to use
a sphere/disk/whatever as a base shape for the light. We consider it to
be a very small geo, always. We could do as Houdini does (and actually
with less code) but I feel its calling for trouble and it doesn't make
sense really. We will wait until someone complains and explain to us
why using a sphere as a spot light is warranted.