WIP for curves (hair) rendering.
We re-use some code from the particles to output a list of attributes.

NOTE: for some reason we are not able to get the correct GT_Primitive
for hair just yet. We get the guides as linear curves (whih we don't want
to have) but not the hair/fur itself.