Fix CMakeLists for macOS build.
We can't use $HIH. It doesn't point to the right installation directory on
macOS and doesn't even exists on Windows it seems.

This commit also uses install() instead of configure_file() to copy the Icon
and the UI files. This is the right way to do it and you have to run:

	ninja install (or make install)

In order to copy all the files properly. Note that the ROP is installed
directly into Houdini because of how Houdini's cmake module is done but I think
that's wrong: ROP should be built in the build directory, and installed using
the install rule.