Fix incomplete merge 126d96949c187fc.
I had already eliminated the m_lights_to_render vector from the context as it
became a simple local variable (since e7a6c644c3f124ea67a4). This is mainly
a simplification as the ROP doesn't need to call scene::find_lights anymore.

As a side effect this change, UpdateLights method in the ROP now takes the
Lights to Render into account, which is good.