Fix UpdateLights (make it works when add or remove lights).
We can't modify the label of an existing PRM_Name. Label is
effectivly modify (if you print it with getLabel) but it has
no effect of what you see on screen!

To deal correctly with adding or removing lights, we have
to destroy all existing multi params and rebuilds it! SAD!