Remove a needles restriction on VOPs exports.
I thought, due to my lack of  experience in Houdini, that VOPs will mostly be
used in a /mat contexts. This is not the case and one can of course define a
"matnet" inside the /obj context. This seems to be the preferred method of
assignment of at least one expert helping with the project. /mat context
is discouraged because of a lack of organisation.

In any case, we now have a better support for material assignment. We don't
support shops yet.