Support VDB volumes rendering through file SOP
For now we limit ourselves to this way of doing things: a file SOP that
references a VDB file in its path. This disallows some workflows such as
instancing of VDB files. As a workaround, note that manual instancing is not
that bad in 3Delight because it will be detected by the renderer and instanced
internally.

Another limitation for now (to be lifted) is a nomenclature restriction for
grids:

  - density
  - temperature
  - heat
  - velocity or vel or simply "v"

Our vdbVolume shader already has a "Grids" tab and we should be supporting
these shortly. Shader assignment is done through the Render->Material
slot.

The implementation skips the GT library refinement for VDBs that are loaded
through the file SOP. This VDB workflow is detected by the
scene::node_is_vdb_loader method and the conversion to NSI can be found
in vdb.{cpp,h}.