Implement object Display flag.
For now, we don't ouput geo that is not displaed. That is a problem if we want
to support this flag in IPR. Another solution could be to output the geo but to
not connect it. Let's see if our users want the display flag to work in IPR
first as exporting usless geo can be wastefulk in Houdini.