Fix prelit/matte when no attribute is present.
When no such attributes are present, we output the objects as both prelit and
matte, which turns objects to invisible basically. I fixed this by checking for
the existence of the parameters on the OBJ (don't forget, in Houdini users
routinely remove attributes).

Note that I don't have attributes because Houdini seems to look for 3Delight.ds
in the  home directory and not the installation directory..