Avoided specifying an empty string as a choice list's item value.
This is the value saved in the scene file, as opposed to the index of the item
(ie : PRM_ORD type).

This is actually good news, since it means that adding new values at any index
won't break older scenes.

However, we still use the index string for shader parameter, since the actual
interface of the shader is really using an integer, and the label is only
additional meta-data. Also, the label is probably more likely to change when the
UI evolves, which should not break existing scenes.