Avoided messing up the CMake configuration with 3DFH traces.
It seems that Houdini's CMake configuration scripts uses hython (Houdini's
Python interpreter), which loads 3DFH (if an old build is laying around in the
search path) which prints stuff that is inappropriate for a makefile.

This is relatively new behavior : it appeared when we switched from Houdini
18.0.246 to 18.0.271 . I don't know if the change comes from the CMake scripts
or from hython, however.