Introduced new "primitive" exporter sub-class.
It serves as a proper base class for exporters of refined GT primitives and
removes related stuff from the exporter base class, where is was unused for
other subclasses light, camera, null and vop.

Most of the GT primitive stuff in exporter was moved to primitive, except for
sm_invalid_gt_primitive, which was completely removed : it became useless
because primitive exporters always have a valid GT primitive.

exporter::export_attributes is only used to export the attributes of GT
primitives right now, so it could be also have been moved to the new primitive
class, but we'll see how Aghiles makes it evolve before doing so.