Export attributes used by dlPrimitiveAttribute VOP
The idea is not to export all the attributes with every single geometry as this
could lead to high memory usage, epecially for particles. Instead, we scan the
VOP network for any dlPrimitiveAttribute shader node and decide what to export
from there.

Note that I opted not to support the bind attribute. There is a log of weird
options in there and we would have to support them all so not to confuse users
with half-baked support. The problem is that the design of that node, used both
for AOVs and attributes, goes against 3Delight philosophy of simplicity an
unambiguity.