Replaced the Export NSI toggle with an option menu.
This removes the need to type "stdout" or an empty string in the file name text
field in order to output the NSI stream to the console. It also simplifies the
labeling.

We used a string option menu with "on" and "off" as 2 of the 3 possible values
so that conversion from an old scene should be trivial, as those are the values
stored in the scene file for a toggle attribute.

Finally, the file name field is now grayed out when unused.