Replaced the Render button with an Export button when the ROP is in export mode.
This makes it clearer that no actual render will occur.

For now, this button's callback simply executes the ROP, which actually exports
instead of rendering because the button is only visible when the ROP is in
export mode. It seems simpler than separating the exporting and rendering logic
in ROP_3Delight.