Passed Python commands into a string argument instead of using a separate file.
$arg0 gets an absolute path when the script is called from Houdini, but a
relative one when it's called from our 456.cmd, so evaluating $arg0.py doesn't
work in that case.

Although it's a bit harder to read, this way of executing the Python code also
has the advantage of not needing an extra file and keeping the patch close to
the problem it fixes.