Avoided assuming that the "3Delight" tab is inserted into the "Misc" tab.
This script is use to patch a bug that inserts our "3Delight" inside an already
existing tab, instead of inserting it at the parent level. However, it should
not assume that it's always inserted into the "Misc" tab, since other plugins
might also add their parameters to the node before us, and it's likely that the
insertion bug simply inserts the 3Delight tab into the last existing tab.