Added 3Delight-specific parameters to environment lights.
Those are similar to those added to regular light sources, apart for the
"spread", which would make no sense on an environment light.
The "prelit" parameter is particularly important for an envlight, since it's the
most likely to be used in a prelit scene.

Since we still have troubles calling an external script from an
automatically-called .cmd script, code from hlight-2.0_OnCreated.cmd was
shamelessly duplicated.