Made the spread parameter visible on "disk" light sources, too.
Also, we simply disable the parameter instead of hiding it when a point light
has "conneenable" off, because it's more likely to be changed during scene
edition, whereas the light type is more likely to be set once and left untouched
afterwards. And we don't like UI that appears and disappears unpredictably.