Changed access to current FPS to avoid binary incompatibilities between versions.
Unfortunately, a member "myDefaultContextOptionMicroNodeLock" was added to
CH_Manager somewhere between Houdini 18.0.287.9 and 18.0.310. This changes the
layout of the class and the behavior of inline function
CH_Manager::getSamplesPerSec (becauses it uses data member "myFPS", which is
located after the new member, so its address offset is not the same for both
versions of Houdini). This breaks our motion blur when the version of Houdini
used to build 3DFH is incompatible with the version that is running.

Using HOM_Module::fps(), instead, fixes the problem because it simply can't be
inlined, as it's virtual and implemented on the application's side.