Stopped relying on Houdini to call our object creation scripts.
In short : it never worked well. See comments in namespace attributes_callbacks
for details.

The creation scripts have been moved to the "scripts/private" directory of the
installation, to avoid them being called automatically. They have been renamed
for consistency, in order to keep attributes_callbacks::add_attributes_to_node()
simple.

Note that we don't need 456.cmd anymore since the callbacks are also called for
each node when a scene is opened or merged. The only tricky part is that we have
to wait until scene loading is finished to actually execute the scripts.