Add support for instancing of multiple primitves.
This can easly be done using, for example, the merge SOP. In which case the
instance receives a number of primtiives. This is a tricky task because we have
a recursive algorithm that refines primitves until we get to a primitive that
NSI understands ... and, we also have a recursive context when we refine the
instanced primitve. To handle this gracefully we keep a separate variable
(m_return) that containss the very last generated primitives. These primitives
can then be tagged as instanced.

Note that motion blurred group of objects still have a bug.