Moved user interface components to the "ui" source directory.
This doesn't change their destination directory, but it avoids hiding them in
multiple sub-directories.

Also, having a properly setup "scripts" directory directly in the source tree is
problematic because, if Houdini is started from the source directory, it will
add it to its search path and automatically executes some scripts directly from
the source tree, even though they're meant to be hidden in the installation's
"scripts/private" directory, and thus ignored by Houdini.