Added "Override Display Flags" toggle to ROP's Scene Elements tab.
This allows us to support Houdini's native display flags by default, while still
retaining the old "Objects to Render" and "Lights to Render" selection patterns
for those who want to use it. It's off by default, so it might break a few
scenes for early users, but it should make more sense to new users.