Avoided closing the static NSI context before the end of an IPR render.
In IPR mode, m_static_nsi and m_nsi share the same handle, but we still don't
want to invalidate it, because it might be needed for later edits.
What we really want to do is close the static NSI output *file* if one has been
opened (which doesn't occur in IPR), so that's what we need to check.