Simplified the logic behind the shaders ramp widgets in order to fix DlToon.
Instead of trying to decode the nomenclature of their parameters (which is
followed very loosely in our shaders), we rely only on their type and the fact
that they share a common prefix, followed by an underscore, to identify them.
Then, we always use their original name directly, in the templates definition.