Exported more attributes to a static NSI file.
The most important of those is instance's "transformationmatrices", as its makes
up for 100% of the exported instances's attributes.  On the other hand,
polygonmesh's "nvertices" is usually not more than 1/12 of the mesh's attributes
(if we estimate 1 "nvertices" count for 4 "P" points).  But it uses some space,
too.

This improves work begun in 5dc288457ab1ec995fcb48daf867d49f67fbfea5 .