Implement primitive SOP materials on instances.
The bonus is that we still use our NSI instance for maximum efficiency. For
that, we create as many "sourcemodels" as there are <material, instance>
pairs.

Note that for the cases where the number of such pairs is extremely high, it is
better to use graph semantics for instancing but this is not the usual use case
for instances.