Implement obj-level instancer
We found a simple way to implement this instancer by using common code with the
SOP-level instancer. This means that we inherit all the hard work we did on the
SOP-level, such as material assignments.

Two things are still needed: rotational transforms and ability to instance
objects that are have the display flag set the off.