Implement "op:" textures.
This means that COP2 nodes are converted to textures and rendered properly.

* In preview/IPR mode, these textures are created in Houdini's temporary
  folder and deleted after each render.
* In NSI export mode, all generated images are saved in the same directory
  as the NSI files but in the /cops subdirectory. In thise case, image
  file names will have the same name as the COP2 node path with the
  frame number appended. Example: img-comp1-OUT-0005.exr
* When exporting NSI to console (stdout) for debugging, nothing is done.

I moved our system functions out of shader_library.cpp and put them in
dl_system name space instead. I think its time we do that. The main reason
for this is that I need some of these functions in cop_utilities.cpp