Implement OBJ-level instancer.
We use the same exporter as for the SOP instancer. We add code to deal with the
"s@instance" attribute and with the various instance orientation attributes (N,
up, v, rot, trans, pivot, scale, pscale and orient).

Implementation had to overflow into scene scanning because the OBJ instancer
can export otherwise invisible objects. So we have do a second run on the
geometries to figure things out. This is done efficiently in near linear time
scan_for_instanced(..).