Remove primitive::get_sop_materials
It's not really needed. We were using it both in the instancer and the polymesh
but that was a mistake. The GT library has already a pretty good API to access
attributes and we really don't need any additional abstraction layer. This also
saves us from potentially copying a large amount of strings. We are talking
about millions of strings here (as many as there are faces in a polymesh).