Don't remove old GEO attributes anymore.
This generates a lot of warnings when uysing Hython. Warnings that we do not
see when sstarting  Maya but that end up flooding logs and users's console that
are using Hython. By now, these old attributes shouldn't exist anyway.

Note that we still get warning on new attribute creation, if the attributes
already exist. We will have to get rid of these too.