Implement Scene Elements -> Matte Objects.
We had non-functional "Matte Sets" in Layers. That's the old school way from
3DFM and I think it makes more sense to have them in Scene Elements. It also
matches the way other renderers, such as Mantra, works.

Note that we still need to think about what to do with "Override Matte
Attributes" and if we need to expose the PreLit tag at this level as well.