Improve scene::convert_to_nsi for nice VOP export
We used to parse scene recursively, looking for objects and vops together,
sometimes getting into nodes we don't want to see. It works, but its a bit
of a mess.

Starting from now, we proceed like this:

1. Scan OBJ nodes (/obj/*)
2. From these objects, gather geometries and get their materials.
3. For each material, export all VOPs, including inputs.

This makes a much cleaner output and find materials that where not even found
before (because we tried to avoid getting into SOPs, and these could contains
VOPs in "matnets"). Now we don't care where the VOPs are hidden as we have the
actual VOP network heads from which we can start our scan.