Renamed accessor exporter::static_attributes_context to attributes_context.
The context returned by this function *might* be the one use for static
attributes, but it's not always the case. Indeed, the role of that accessor is
to choose which context to return.
The defining characteristic of the returned context is, that it's meant to
export the object's  attributes, so the accessor should be named as such.