Fix visiblity bug introduced with OBJ instancer.
The visiblity became hierarchical: if the parent was not visible the children
would be invisible too. This hapenned because I decided to "clean" the null
transforms of invisible nodes. This is was required for the instancer
implementation.

Note that the good visiblity behaviour was working by "mistake" as was not
something we paid attention to. At least know we know why we need to output
a null for each object, even if invisble. :)