Ensured that OBJ nodes nested in objnets also receive 3Delight attributes.
For instance we can now detect the creation of OBJ nodes
/obj/chopnet1/objnet1/sphere1 or /out/objnet1/torus1 .

This implies adding callbacks to any "manager" type node in order to detect the
creation of all OBJ-manager nodes, to which we also add a callback to detect the
creation of OBJ nodes. It looks like lots of callbacks but, in the end, we only
add those to "manager" type nodes.