Change some parameter types and fix shader type.
This is needed to have networks connect to the correct terminal type in the USD
universe. For shaders without a closure output, the shader type is used. For
those with a closure output, it is the type of that parameter which is used.

This also makes a distinction between color and closure parameters, such that
connecting them together now results in a big warning in Houdini. This is good
because it won't work.