Moved handling of bundle-specified mattes to geometry exporter.
There is no need to scan the whole scene to find "bundled" matte objects since
we already have a bunch of exporters that can handle this by themselves.

This is especially important in IPR mode, where we won't want to re-export *all*
matte connections each time an object is re-exported.