Removed the 3Delight Material Builder from the export logic in "scene" class.
There is no exporter for the Material Builder, and we actually don't need one
since it has no equivalence in NSI.  This is one less thing to worry about.

This avoids exporting the builder's "main" material and its connections twice
(once with the builder's handle, once with the actual shader's handle). This
caused changes to the material to go unnoticed in IPR because, while the
"material" shader was updated in the NSI scene, NSI objects were actually
connected to the "builder" shader.
It also eliminates the VOP's "Am i just a material builder? And do i even refer
to a valid object?" identity crisis introduced in
2a1a43a911b7607b8261509f22a4d295a661f565 . In particular, it eliminates the
comparison of the operator's name with incorrect string "dlMaterialBuilder",
which should have been updated in 3120b97c4abf72fcf071194e66f87ffbfc3c0376 .