Added support for geometry edits in IPR render.
The tricky part here is catching SOP-level changes, which don't trigger the
callback attached to the OBJ node. Fortunately, they seem to be easily detected
by connecting a single callback to the OBJ's render SOP.

Without this SOP-level callback, even something as simple as creating a new
sphere wouldn't work because as soon as the sphere is created, its SOP's radius
is divided by 2, and that would go unnoticed.