Fix custom AOV scanning.
We now take into consideration objects visibility when scanning for
AOVs. Also, we don't scan just any VOP in the scene, only these
that are somehow connected to rendered materials.

There is some code re-organisation so that I can reuse VOP scanning
and object visiility logic. I decided to make a new class for the
later as I feel we might need this again.