Fixing the incandescence light.
I had to do much more work in the end as we had many structural issues in the
multi-light exporter. Some notes:

* We remove the "null" transform above the "light" as we will be re-using that
  handle to create an NSI "set" node to which we can attach the objects. We need
  a handle that points to the right place in the scene and the "null" handle
  was in the way (and not needed anyway).
* Incadescence lights cannot be bundled. It doesn't make sense anyway as it is
  already a group of objects.

There is still some work to be done to iron out some final details. I see that
hiding the incand light will create a layer with RGBA in it for example.