Removed output path from exporter::resolve_material_path.
It's seldom used since fd63817026592d614c1d717e1bcd26ba56f258d6 , and it's easy
to retrieve the material's full path afterwards if that's actually what is
needed.
That way, the function performs only one tasks, and there is less temptation to
use the material's full path in ways that are inappropriate, such as an NSI
handle.