Started using the node's "unique ID" as a NSI handle in IPR. (fixes #132)
This fixes issue 132, which was caused by two nodes (one being still incomplete)
sharing the same full path during a copy & paste operation.

It also improves the similar issue 131 (by preventing the deletion of the source
node), but this still doesn't fixes it completely (as the pasted node is still
not exported once it has finished initializing).