Fix and accelerate scanning of bind attributes.
Those are the attributes accessed by nodes like dlAttributeRead. The fix
concerns the case where both SOP-level and OBJ-level assignents are used
(This is possible by leaving empty stirngs in @shop_materialpath). In that
case we were only exporting attributes needed for te SOP-level part and not
for the OBJ-level part.

Regarding the acceleration: Houdini has an indexed string representation
that allows de-duplicting primitive string variables. So if you have 1000
times the same string, you can access that 1 string without iterating
throught the entire array. So we use that functionality and it should save
significant time, as we have seen in aa943786dd6f8e.