Fix instancing in IPR when source geo is invisible
A mixup between what should be a handle and what should be a path to
an object is at the root of the problem. Since 8d6744dc9161e9.

instance::get_instanced was not good as well as it returned handles
in the SOP-level instancer ... but these handles are internal to
the SOP and shouldn't be returned at all.

Note that instancing also concerns geo lights, whish were broken
as well when source geo is invisible (same code line).