Added an Export to Archive render mode.
It creates an NSI archive file meant to be read by a procedural node.
This file excludes global options, output layers and drivers, as well as the
atmosphere shader.  It also doesn't refer to a "static" NSI file for
non-animated objects, in order to keep the archive as self-contained as
possible.