Added support for current time change in IPR. (fixes #143)
Unfortunately, changing the application's current time doesn't result in the
proper parameter change events being sent to our node interests callbacks. This
means that we have to react explicitly to time change events and re-export the
nodes that need it.

When a time change occurs, we delete the NSI nodes associated to all
time-dependent Houdini nodes, then we re-export the time-dependent elements of
the scene. Since we don't export invisible objects and their visibility may not
be the same as in the previous frame, the nodes we delete and those we export
might not be exactly the same. This means that we can't simply rely on the
exporters to delete their previous representation, because that would result in
newly hidden objects remaining visible.

The template "deleter" class allows the logic behind the decision to create an
exporter (or not) to remain in process_obj_node, and the NSI calls to still be
deferred after all exporters have been created.