Support s@instancefile with NSI standins in OBJ instancer.
Since we can now export NSI Stand-Ins, it is useful to be able to use
s@instancefile with the OBJ-level instancer. Note that SOP-level instancing
is already possible but the end result is different:

OBJ-level: 3Delight creates procedural that loads the NSI stand-in.
SOP-level: The object is inside the main NSI context (file) and is
referenced as an intancer.

In the end its a matter of taste for the users as both ways will yield
to same performance unless there is *alot* of different procedurals to load,
in which case the multi-threaded loading of procedurals will win for the
OBJ-level instancer.