Fixed File Output when OpenEXR (deep) Image Format is used. Fixes #151
When selecting OpenEXR (deep) from Image Format dropdown this would
create a file with .deepexr extension, something that cannot be read in
Nuke and does not exist. What we changed is only the file output extension
from .deepexr to .exr whereas the drivername would remain the same (in this
case deepexr).