Fixed export to NSI for array type parameters (int[],float[]) Fixes #175
Some parameters such as depth, on cloud and marble shader were exported
as float instead of float[2] (min and max). This was happening because
when exporting NSITypeInteger and NISTypeFloat we were not taking in
consideration the array length of the parameter, thus exporting only the
first parameter in case of an array.