Added Radius Paramter on SpotLight to control it's shadow. Fixes #179.
By default the radius of the circle that emits light was very small (0.001),
making the cone almost ideal but in the meantime producing very sharp lighting.
So to soften the lit region edge and the shadow casted by spot light we
added a Radius parameter which will control this value.