Don't use background render for networked rops.
When a 3Delight ROP has an output connected, we wait for the render to be
finished insted of starting it in a background thread. This allows for
sequential rendering of said ROPs.

While at it, removed m_rop_path from public members as the thing we need is the
ROP, not the path. It's was just a bit silly to use.