Add support for heightfields rendering.
This is done, as usual, through the GT API. We receive a converted polygon
mesh. I am not 100% sure what variables come attached to it. Trying Cd seems
to give something but not close to the terrain color one sees in the viewer.

It's only enabled for builds higher than 18.0.532. Note that this API
has been backported from 18.5 branch of Houdini.