Added Closure data types on UI as Color but disabled.
Closure data types are now part of the UI (except of aovGroup) since it
is defined as closure on osl shader, but we do not show it. Since they do
not affect the surface using a color value we are disabling those parameters
and leaving the possibility to connect them using the VOPFX menu.