Added pythonX.Xlibs to our plugin so it recognizes imported python files
Except from adding a new feature on our plugin, this commit would fix
Issue #199 as well.As mentioned on the commit c5fc01aa Houdini would
only recognize python files if they were located on
$Houdini_Path/houdini/python2.7libs. But as seen from the documentation
the python2.7libs would be added to the Python path if it was in any
directory on $Houdini_Path so we are adding it to our plugin installation
directory. Not the items on 3Delight menu on of VOPFx menu will be
added without the need of the python file being located outside of our
plugin directory.