Rewrited code for the added invisible parameters to avoid errors.
Since we updated to new ROP UI, some of the parameters were not being
used on each of the ROP, and to avoid the error regarding evaluating
these unknown parameters, we had added them as invisible on the ROP's UI
from where they were missing. We have writen the code in a nicer way and
added comments explaining the changes made.