Automatically connect makextransform node with 3D Textures input node.
Since there are many nodes that output matrices, we do not want the
users to get confused and guess which one is supported by us. To avoid
this confusion we are automatically creating and connecting our transform
matrix node (makextransform) to the placementmatrix input parameter of 3D
Patterns. So this transform matrix node will only be created when we
created a 3D Pattern shaders from 3Delight shaders.