Use makexform osl shader instead of place3dTexture osl shader.
Instead of checking everytime for makexform node and then replace
it with palce3dTexture to get the proper osl file, we are adding a new
osl shader to our plugin to avoid this. Also this would avoid adding an
extra parameter to our pre-existent osl placement matrix, in order to
make it part of our shaders group.