WIP: Initial dlAOVGroup implementation.
Implemented the UI for our dlAOVGroup VOP_Node. The user can choose the
number of the AOVs using the incremental button which also increases the
number of the connections in the node network. Also when changing the
name of the AOVs from the string input, the connection name (input
label) will change too.

TODO: The AOV inputs created should be static on all the dlAOVGroup nodes
created. This means that once we create several AOVs for one dlAOVGroup,
those AOVs should be present on all other dlAOVGroup created too.