Updated the way we calculate placementMatrix output.
Instead of building the matrix on c++ file we built it on our osl
shader, and calculated the output matrix there. This would make
connections to the makexform VOP_Node work, and also update IPR without
having any additional code for it.
For now we only support translate, rotation and scale. Transform order
and Rotation order are not yet supported.