Finished dlAOVGroup connection with surfaces and it's selections on ROP.
Now the user can use the dlAOVGroup instead of bind export node in order
to create custom aov layers. Each dlAOVGroup can have more then one
layer and after we connect the shader with the dlAOVGroup on it, we can
select from the ROP which of those custom aovs we want to render.