Getting shutter time value from viewport active camera.
In order to achieve this we have renamed viewport_hook class to
dm_SetupTask becuase it is a friend class of DM_VPortAgent class and
makes it possible to access it's private members, which in our case
it is the active camera on viewport which we get. Most of the camera
attributes are found on GUI_ViewParameter, which we have already used,
but the shutter time parameter is not part of it, so we are using the
OBJ_Camera to get that value.