Fixed connection to AOVGroup shader.
Connecting other shaders to AOVGroup to create a custom layer would only
work if the outputLabel of the connected shader was outColor. This
woould render fine for most of the shaders since they have the same
outputLabel name but some others would give an error because of their
outputLabel being different. Now this will work for any outputLabel
connection that you have.