Getting the correct value of active camera's shutter duration on viewport.
When rendering on Houdini's viewport you can render multiple viewport's
at the same time, thus we need to get the right shutter value for each
of them. This also fixes and removes the need of having an additional hidden
camera which was not a good solution at all (always the first created
camera was taken into account for motion blur).