Fixed incandescence light AOV problem. Fixes #299
In the initial fix we did not consider that the display value might
collide when overriding display flags. This issue has now been fixed,
Also fixed the error message appearing when having the incan light
on the scene but not the object it corresponds. This error would be as a
result of trying to set incanlight_multiplied to an unexistent shader
node, since the object to where it's connected does not exist on the
scene.