Updated Matte Objects in ROP to only require objects of that list. Fixes #314
If we were to override the Objects to Render list, and have several
objects on Matte Objects list, then we should have added these objects
on "objects to render list" too, in order to render them as mattes.
Now objects that are added to Matte Objects are automatically rendered
without needing to add them on Objects to Render list.