Fixed custom AOV issue when connecting multiple outputs to the same AOV. Fixes #337
AOV group was not working correctly when having multiple inputs on the
same AOV node. This because we were not getting the correct output label
connected to the dlAOV's input node. This would also fix the crash
mentioned on the fixed issue, where the input's number connected of the
AOV node is greater than the connected node's output number.