Fixed error messages when creating a light source during IPR.
Since light sources are also cameras in houdini, creating lights during
IPR would try to set non-existent parameters that are part of camera
objects only. These parameters are evaluated on get_screen_window()
function and we were not encountering this problem before because this
function was being called on ROP_3Delight.cpp and not on camera.cpp.