Simplified viewport resolution change logic. (fixes #374)
The buffer is now initialized completely upon construction. Also, it's not
disconnected when the resolution change is detected on the viewport_hook
rendering side, anymore (ie : in viewport_hook::export_screen_attributes).
Instead, we wait for the display driver to need a buffer with a different
resolution to re-allocate one.

This avoids disconnecting the buffer from the viewport too soon, which seemed
to cause problem when buckets were received around the same time.