Deal with aborting from Viewport ROP shelf command without errors.
Since there is only one shelf button which handles both viewport start
and abort (based on the viewport render situation), trying to abort it
would cause error of invalid output driver, since it tries to re-render
although it has been disconnected. To solve this we do a check for all
viewport ROP, if any of them is currently active (rendering) we stop
them as we have pressed Abort button.