Fix custom aovs crash when corresponding node display flag is off. Fixes #388
If display flag of a node whose connected material is being used as a
custom AOV layer is off, then this would result in crush after the
scene is opened again and rendered. This because we were updating the
AOVs before the node's parameters were completed when we were opening
a previously saved scene. This commit will also avoid the crash if an
aov connection is removed from dlAOVGroup, but the aov list on ROP has
not been updated yet.