Rattle Houdini a little before asking if an invisible object is time-dependent (take 2).
Otherwise, it would wake up, too late, during the export process and the first
frame will have incorrect motion blur.

This seems to fix issue 268.

This is the second try for this patch.  The original one was made in
0c413cf169ac98ab4f63197026bd89b0dcec22a1 and
3ef00d52fbd7f09af6f289ef587200497a66ef8d, then reverted in
86f534b7366f51cdb874b62a15aa3ec987a27da8 and
46eff7c269b9354933fa912a3188ea34660710b4 .  The current commit re-applies the
original patch, with an extra condition in time_sampler::is_time_dependent which
checks whether we actually want to render the object before "rattling" it.