Support all kind of outputs when debugging node with debug flag. Fixes #407
Previous debug flag implementation was assuming that all shader's
outputs were named outColor, leading to wrong behavior when having
different output name. Except for fixing this, we are also giving
priority to the connected output to debug and then the color output.
E.g for the float output, we would use that to modify specular weight,
rougness, weight etc, and would like to see how it looks.